Added plugin option
