+++ src/plugin.ts
@@ -1,0 +1,17 @@
+import { App } from 'vue'
+import { ContextStateSymbol, ContextUpdateSymbol } from './context'
+import useUnleashProvide, { IProvideOptions } from './useUnleashProvide'
+
+
+const plugin = {
+  install(app: App, { config, unleashClient, startClient }: IProvideOptions) {
+    const { context, update, start } = useUnleashProvide({ config, unleashClient, startClient })
+
+    app.provide(ContextStateSymbol, context)
+    app.provide(ContextUpdateSymbol, update)
+
+    start()
+  }
+}
+
+export default plugin
